+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# NEXT
+
+- feat: set scroll position to currently active slide on navigation drawer open
+
 # 0.10.1
 
 - refactor: rework flutter deck controls